// src/app/model/basar-model.ts
export interface  Customer {
  basarNumber: number;
  name: string;
  street: string;
  houseNumber: string;
  postalCode: string;
  city: string;
  phonenumber: string;
  email: string;
}

export interface Product {
  lfdNo: number;
  category: string;
  description: string;
  size: string;
  price: string;
  basarNumber: string;
}